requirements.txt
streamlit==1.29.0
anthropic==0.52.0
mistralai==0.2.0
pymupdf==1.24.0
pillow==10.3.0
pandas==2.2.0
reportlab==4.0.0
boto3==1.34.40
botocore==1.34.40
uuid==1.30
llama-cloud-services
python-dotenv